requirements.txt
flask
flask_bootstrap
newsapi-python